Crate cryptocompare

General

An rust API-Wrapper for the CryptoCompare API.

Link to www.cryptocompare.com/api/ API.

Structs

Dataset
An object with trading info.

Options
Informations for the request.

Functions

get_coin_list
Get general info for all the coins available via CryptoCompare.

get_history_day
Get open, high, low, close, volumefrom and volumeto daily historical data.

get_history_hour
Get open, high, low, close, volumefrom and volumeto from the each hour historical data.

get_history_minute
Get open, high, low, close, volumefrom and volumeto from the each minute historical data.

get_price
Get the latest price for a list of one or more currencies.

get_price_historical
Get the price of any cryptocurrency in any other currency that you need at a given timestamp.

get_price_multi
Get a matrix of currency prices.

get_price_multi_full
Get all the current trading info of any list of cryptocurrencies in any other currency that you need.

parse_json_to_vector
Helper function to parse a JSON-Value to a Vector of Datasets